content: net: Add VIRTIO_NET_F_HASH_OFFLOAD feature
This commit introduces hash offload support into network device. The device
calculates hashes for packets all the time and especially when it uses
multiqueue, these hash values can be very useful for the driver and OS in some
cases.

This feature allows the delivery of the calculated hash to the driver.

Signed-off-by: Sameeh Jubran <sjubran@redhat.com>